requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.0
pydantic==2.9.0
groq==0.11.0
sentence-transformers==3.1.1
PyJWT==2.8.0
bcrypt==4.1.1
